Re-organised code
Closes #663
